Consolidate and organize file operations
